requirements.txt
langgraph>=0.2.60
streamlit>=1.41.1
langchain-google-genai>=2.0.7
python-dotenv>=1.0.1